requirements.txt
asgiref==3.9.1
backports-datetime-fromisoformat==2.0.3
Django==4.2.23
djangorestframework==3.16.0
environs==14.2.0
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.4.0
inflection==0.5.1
iniconfig==2.1.0
marshmallow==4.0.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
pytest-factoryboy==2.8.1
python-dotenv==1.1.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
